requirements.txt
openai
python-dotenv
beautifulsoup4
requests
flask
flask-cors
markdown
gunicorn